Module3_Databases_and_REST_APIs/Lessons/Lesson5_Building_RESTful_APIs/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
mysql-connector-python
marshmallow-sqlalchemy



# Command to install the packages
pip install Flask Flask-SQLAlchemy Flask-Marshmallow mysql-connector-python marshmallow-sqlalchemy
